+++ app/Http/Controllers/Admin/ProductController.php
@@ -6,20 +6,24 @@ use App\Http\Controllers\Controller;
 use App\Models\Category;
 use App\Models\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 
 
 class ProductController extends Controller
 {
-    
+
     public function index()
     {
+        $vendorId = Auth::id();
+        $products = Product::where('user_id', $vendorId)->get();
+
         return view('admin.products.products', [
-            'products' => Product::all()
+            'products' => $products
         ]);
     }
 
-   
+
     public function create()
     {
         return view('admin.products.create', [
@@ -27,21 +31,20 @@ class ProductController extends Controller
         ]);
     }
 
-   
+
     public function store(Request $request)
     {
         $validated = $request->validate([
             'title' => 'required|max:255',
-            'image' => 'required|image|max:50',
+            'image' => 'required|image|max:1024',
             'about' => 'required|max:1000',
             'price' => 'required|numeric|min:0',
             'stock_quantity' => 'required|numeric|min:0',
             'discount' => 'multiple_of:5|min:0',
             'category_id' => 'required|exists:categories,id'
         ]);
-
-        $path = Storage::putFile('products', $validated['image']);
-
+        $vendorId = Auth::id();
+        $path = $request->file('image')->store('products', 'public');
         $product = new Product();
         $category = Category::find($validated['category_id']);
 
@@ -50,8 +53,9 @@ class ProductController extends Controller
         $product->about = $validated['about'];
         $product->price = $validated['price'];
         $product->stock_quantity = $validated['stock_quantity'];
+        $product->user_id = $vendorId;
 
-        $request->whenHas('discount', function ($input) use ($product){
+        $request->whenHas('discount', function ($input) use ($product) {
             $product->discount = $input;
         });
 
@@ -60,11 +64,11 @@ class ProductController extends Controller
         return redirect()->route('admin.products.index')->with('status', 'product added successfully');
     }
 
-    
-    // public function show(Product $product)
-    
 
-    
+    // public function show(Product $product)
+
+
+
     public function edit(Product $product)
     {
         return view('admin.products.edit', [
@@ -73,24 +77,26 @@ class ProductController extends Controller
         ]);
     }
 
-    
+
     public function update(Request $request, Product $product)
     {
         $validated = $request->validate([
             'title' => 'required|max:255',
-            'image' => 'required|image|max:50',
+            'image' => ['image', 'mimes:jpeg,png,jpg,gif', 'max:2048'],
             'about' => 'required|max:1000',
             'price' => 'required|numeric|min:0',
             'stock_quantity' => 'required|numeric|min:0',
             'discount' => 'multiple_of:5|min:0',
             'category_id' => 'required|exists:categories,id'
         ]);
-        
+
         $product->title = $validated['title'];
 
-        if($request->has(['image', 'discount'])){
-            Storage::delete($product->image);
-            $path = Storage::putFile('products', $validated['image']);
+        if ($request->has(['image', 'discount'])) {
+            $path = $request->file('image')->store('products', 'public');
+            if ($product->image) {
+                Storage::delete($product->image);
+            }
             $product->image = $path;
 
             $product->discount = $validated['discount'];
@@ -108,13 +114,13 @@ class ProductController extends Controller
         return redirect()->route('admin.products.index')->with('status', 'product updated successfully');
     }
 
-   
+
     public function destroy(Product $product)
     {
         Storage::delete($product->image);
         $title = $product->title;
         $product->delete();
 
-        return redirect()->route('admin.products.index')->with('status', 'product "'.$title.'" deleted successfully');
+        return redirect()->route('admin.products.index')->with('status', 'product "' . $title . '" deleted successfully');
     }
 }
